requirements.txt
PyYAML >= 6.0.1
yamlable >= 1.1.1
glfw >= 2.6.2
imgui-bundle >= 1.0.0
